requirements.txt
fastapi
jinja2
pydantic
python-escpos
python-multipart
pyusb
requests
sqlalchemy
tabulate
tomli
tomli_w
uvicorn[standard]
websockets
